ADD: All yaml files and demo.sh script
demo.sh may be used to run the demo yourself.
It uses kubectl command and the yaml files found in this repository.
Purpose of the yaml files is to show how to describe deployments,
services and rules in kubernetes and make both HAProxy and Brocade vTM
happy with the live changes happening there.